% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eap.R
\name{eap}
\alias{eap}
\alias{eap,ANY-method}
\title{Finding EAP}
\usage{
eap(raschObj, lower = -6, upper = 6, ...)
}
\arguments{
\item{raschObj}{An objet of class Rasch}

\item{lower}{A numeric value, the lower limits of integration (defaulted to -6)}

\item{upper}{A numeric value, the upper limits of integration (defaulted to 6)}
}
\value{
\item{ans}{The expected a posteriori value for thetaJ calculated using equation (3)}
}
\description{
Finds the expected a posteriori value for thetaJ using equation (3)
}
\note{
This is a very simple function
}
\examples{
eap(testRasch,-5,5)
}
\seealso{
\code{\link{print}}
}
\author{
Gangyi Sun
}
