% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAPFunction.R
\name{EAPFunction}
\alias{EAPFunction}
\title{Calculates Expected A posteriori value for proposed theta}
\usage{
EAPFunction(raschObj = "Rasch", lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{an object containing a students name, ability and whether he got a question correct}

\item{lower}{lower bound of integration; defaults to -6 if not specified}

\item{upper}{upper bound of integration; defaults to 6 if not specified}
}
\value{
Expected value of theta
}
\description{
Calcualtes Expected A posteriori value using Rasch model, as a fucntion of the RaschObject (data)
}
\note{
Calls likelihood funcoitn and priorProbFunction
}
\examples{

testRaschObj = new("Rasch", name="Jon",a =c(2,4,3), y =c(1,1,0))
EAPFunction(test1,-5,5)
}
\author{
Jonathan Gross <\email{jonathan.gross@wustl.edu}>
}
