% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAPRasch}
\alias{EAPRasch}
\alias{EAPRasch,ANY-method}
\title{Calculate the expected a posteriori value}
\usage{
EAPRasch(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limites of integration}

\item{upper}{The upper limites of integration}
}
\value{
The output containing
 \item{theta_hat}{expected a posteriori value for theta}
}
\description{
Calculate the expected a posteriori value for theta, given the likelihood function and the prior
function
}
\note{
This helps to work in a Bayesian framework
}
\examples{

Trump <- new("Rasch", name = "Trump", a = -3:3, y = c(1, 1, 0, 0, 1, 0, 1))
EAPRasch(Trump, -5, 5)
}
\author{
Luwei Ying
}

