% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch-package.R
\docType{package}
\name{easyRasch}
\alias{easyRasch}
\alias{easyRasch-package}
\title{easyRasch}
\description{
The easyRasch package provides an S4 class and methods for Rasch models.
}
\author{
JB Duck-Mayr
}
