% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherInf.R
\name{fisherInf}
\alias{fisherInf}
\title{Calculate the Fisher's Information for a Rasch model}
\usage{
fisherInf(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{theta}{A proposed value of theta}
}
\value{
A vector of Fisher's Information values corresponding to the number of items
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 exRasch <- new('Rasch', name = 'Me', a = 0, y = 1)
 fisherInf(exRasch, 0)
 @seealso eap
 @rdname fisherInf
 @aliases fisherInf,ANY-method
 @export
}
\description{
Calculates the observed Fisher's Information for each question in a test given
the test taker's ability parameter (theta) from a \code{Rasch} object.
}
