% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoCurve.R
\name{InfoCurve}
\alias{InfoCurve}
\alias{InfoCurve,ANY-method}
\title{Plot the information curve}
\usage{
InfoCurve(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limit of integration}

\item{upper}{The upper limit of integration}
}
\description{
Plot a test information curve to see how much information we gain about test takers 
for a broad range of abilities.
}
\note{
This function is created to visualize the Fisher's Information
}
\examples{

Trump <- new("Rasch", name = "Trump", a = -3:3, y = c(1, 1, 0, 0, 1, 0, 1))
InfoCurve(Trump)
}
\author{
Luwei Ying
}

