% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfun.R
\name{lfun}
\alias{lfun}
\alias{lfun,ANY-method}
\title{Likelihood Function}
\usage{
lfun(RaschObj, theta)
}
\arguments{
\item{RaschObj}{An object of "Rasch" class}

\item{theta}{A proposed value of theta}
}
\value{
a vector
\item{}{The likelihood value}
}
\description{
The function calculating the Rasch likelihood
}
\note{
This function is just for the exam.
}
\examples{
atesttaker<-new("Rasch", a=c(1:10), yj=rep(c(1,0),5))
lfun(atesttaker,2)
}
\author{
Jaerin Kim
}
