% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Like.R
\name{Like}
\alias{Like}
\title{Calculating Theta Likelihood}
\usage{
Like(raschObj, theta)
}
\value{
A numeric representing the likelihood theta is correct
 \item{prod(PQ)}{The product of vector output PQ}
}
\description{
Calculating the Likelihood That a Proposed Value of Theta is Correct
}
\section{Slots}{

\describe{
\item{\code{raschObj}}{An object of class Rasch}

\item{\code{theta}}{A numerical input for theta}
}}

\examples{

myRasch <- newRasch("Emily", c(3, 4, 12), c(1, 1, 0))
Like(myRasch, theta = 4)
}
\seealso{
newRasch
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
