% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\alias{likelihood,ANY-method}
\title{Finding likelihood}
\usage{
likelihood(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An objet of class Rasch}

\item{theta}{A numeric value, the proposed value of ThetaJ}
}
\value{
\item{like}{the calculated likelihood from equation(2)}
}
\description{
Finds the likelihood of a proposed value of thetaJ
}
\note{
This is a very simple function
}
\examples{
qDiff<-c(1,2,3,4,5)
qAns<-c(1,1,0,1,0)
testRasch<-new("Rasch", testTakerName="Jim", a=qDiff, y=qAns)
likelihood(testRasch,3)
}
\seealso{
\code{\link{print}}
}
\author{
Gangyi Sun
}
