% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunction.R
\name{likelihoodFunction}
\alias{likelihoodFunction}
\title{Calculates Likelihood of proposed theta value}
\usage{
likelihoodFunction(raschObj = "Rasch", theta = "numeric")
}
\arguments{
\item{raschObj}{an object containing a students name, ability and whether he got a question correct}

\item{theta}{a numeric value representing a propsed theta value}
}
\value{
The likelihood -- the probablitiy that the proposed theta value is correct given the data
}
\description{
Calcualtes the like likelihood using rasch model of a proposed theta value, using product of probablities
}
\note{
Calls probFunction
}
\examples{

testRaschObj = new("Rasch", name="Jon",a =c(2,4,3), y =c(1,1,0))
theta = 3.5
likelihoodFunction(test1,theta)
}
\author{
Jonathan Gross <\email{jonathan.gross@wustl.edu}>
}
