% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{LikelihoodRasch}
\alias{LikelihoodRasch}
\alias{LikelihoodRasch,ANY-method}
\title{Calculate the likelihood}
\usage{
LikelihoodRasch(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of theta.}
}
\value{
The output containing
 \item{Likelihood}{The joint probability of all of the observed responses given our guess of theta}
}
\description{
Calculate the ``likelihood" of a proposed value, theta, given the responses we have 
observed from the test taker
}
\note{
This is to calculate the ``likelihood" of a proposed value
}
\examples{

Trump <- new("Rasch", name = "Trump", a = -3:3, y = c(1, 1, 0, 0, 1, 0, 1))
theta <- 1
LikelihoodRasch(Trump, theta)
}
\author{
Luwei Ying
}
\seealso{
\code{\link{Probability}}
}

