% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liklihood.R
\name{Liklihood}
\alias{Liklihood}
\title{Runs Liklihood function}
\usage{
Liklihood(raschObj = "Rasch", theta = "numeric")
}
\arguments{
\item{raschObj}{a object of class Rasch}

\item{theta}{a numeric designating theta}
}
\value{
A numeric that is the product of PQ values
}
\description{
Runs Liklihood function
}
\note{
I am not exactly sure what math I am suppose to do here. This is my best intepretation
}
\examples{

}
\author{
Jack Ploshnick
}
