% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llRasch.R
\name{llRasch}
\alias{ANY-method}
\alias{llRasch}
\alias{llRasch,}
\title{Likelihood of a proposed value of theta}
\usage{
llRasch(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of \eqn{\theta_{j}}}
}
\value{
The calculate likelihood for a proposed theta given y
}
\description{
Likelihood of a proposed value of theta
}
\examples{
set.seed(1235)
Maggie <- new("Rasch", name = "Maggie", 
a = sample(-3:3, size = 7, replace = TRUE),
y = sample(c(0,1), size = 7, replace = TRUE))
llRasch(raschObj = Maggie, theta = 2)
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\seealso{
\code{\link{Rasch}}
}

