% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{Create new Rasch Object}
\value{
A list with the elements
 \item{x}{The first object input} 
 \item{y}{The second object input}
 \item{z}{The third object input}
}
\description{
Creates a new object of S4 class Rasch
}
\section{Slots}{

\describe{
\item{\code{name}}{A character object indicating the name of the test taker}

\item{\code{a}}{A numeric object indicating a vector of question difficulty parameters}

\item{\code{y}}{a A numeric object with the same dimensionality as \code{a} comprised of zeroes and ones indicating a vector of answers for the respondent}
}}

\examples{

myA <- c(3, 12, 3) 
myY <- c(1, 1, 0) 
newRasch <- ("Emily", myA, myY)
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
