% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfun.R
\docType{methods}
\name{initialize,Rasch-method}
\alias{initialize,Rasch-method}
\alias{pfun,ANY-method}
\title{Probability Function}
\usage{
\S4method{initialize}{Rasch}(.Object, ...)
}
\arguments{
\item{RaschObj}{An object of "Rasch" class}

\item{theta}{A proposed value of theta}
}
\value{
A list of two vectors
\item{pr}{The probability function}
\item{pq}{The probability of yj taking the value given}
}
\description{
The function calculating the Rasch probability
}
\note{
This function is just for the exam.
}
\examples{
atesttaker<-new("Rasch", a=c(1:10), yj=rep(c(1,0),5))
pfun(atesttaker,2)
}
\author{
Jaerin Kim
}
