% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPlot.R
\docType{methods}
\name{plot,Rasch-method}
\alias{plot,Rasch-method}
\title{Printing a Rasch object}
\usage{
\S4method{plot}{Rasch}(x, EAP = TRUE, together = TRUE)
}
\arguments{
\item{x}{An object of class Rasch}

\item{EAP}{Should a dashed line appear showing the EAP for \eqn{\theta_j}?}

\item{together}{Should the item characteristic curves be drawn in one
plotting panel, or should they be drawn in separate panels using lattice
graphics?}
}
\description{
For an object of class \code{Rasch}, prints the item characteristic curves.
The x axis is possible values of \eqn{\theta_j}; the y axis is \eqn{P_ij}.
}
\seealso{
\code{\link{Rasch-class}}
}
\author{
JB Duck-Mayr
}
