% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFisherRasch.R
\name{plotFisherRasch}
\alias{ANY-method}
\alias{plotFisherRasch}
\alias{plotFisherRasch,}
\title{Test Fisher Information Plot}
\usage{
plotFisherRasch(raschObj, minTheta = -5, maxTheta = 5, title = NULL, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{minTheta}{A minimum value for theta.}

\item{maxTheta}{A maximum value for theta.}

\item{title}{A character vector with the title defined by the user. If not specified,
the title will be "Information Plot for 'name'".}
}
\value{
A plot with sum of the Fisher Information for a given theta in the y-axis
and the values of theta in the x-axis.
#' @author Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\description{
Test Fisher Information Plot
}
\examples{
set.seed(123)
Maggie <- new("Rasch", name = "Maggie", 
a = sample(-3:3, size = 16, replace = TRUE),
y = sample(c(0,1), size = 16, replace = TRUE))
plotFisherRasch(raschObj = Maggie, minTheta = -10, maxTheta = 10)
}
\seealso{
\code{\link{Rasch}}
}

