% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{PlotRasch}
\alias{PlotRasch}
\alias{PlotRasch,ANY-method}
\title{Plot for Rasch class}
\usage{
PlotRasch(raschObj, lower = -6, upper = 6, EAP = TRUE)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limit of integration}

\item{upper}{The upper limit of integration}

\item{EAP}{whether to have EAP added to the plot}
}
\description{
Take in a Rasch object and create a plot with several parameters
}
\note{
This function is created to visualize the Rasch functions
}
\examples{

Trump <- new("Rasch", name = "Trump", a = -3:3, y = c(1, 1, 0, 0, 1, 0, 1))
PlotRasch(Trump)
}
\author{
Luwei Ying
}

