% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printe.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\title{Rasch Method for Print}
\usage{
\S4method{print}{Rasch}(x)
}
\arguments{
\item{x}{An object of "Rasch" class}
}
\value{
Name and expected a posteriori value of the object
}
\description{
Prints expectation of a posteriori.
}
\note{
This function is just for the exam.
}
\examples{
atesttaker<-new("Rasch", a=c(1:10), yj=rep(c(1,0),5))
print(atesttaker)
}
\author{
Jaerin Kim
}
