% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.raschObj.R
\name{print.raschObj}
\alias{print.raschObj}
\title{Prints name of the test taker and the EAP results}
\usage{
print.raschObj(raschObj)
}
\arguments{
\item{raschObj}{An object of class Rasch}
}
\value{
A list with the elements
 \item{name}{Name of the test taker}
 \item{eap}{The EAP results}
}
\description{
Returns name of the test taker and the EAP results
}
\author{
Rohan Gupta<\email{rohangupta@wustl.edu}>
}
