% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\alias{print,ANY-method}
\title{Print}
\usage{
\S4method{print}{Rasch}(x, ...)
}
\arguments{
\item{raschObj}{An objet of class Rasch}
}
\value{
A list containing:
 \item{name}{The name of the test taker}
 \item{eapResult}{The EAP results of raschObj}
}
\description{
Prints name of test taker the expected a posteriori value for thetaJ using equation (3)
}
\note{
This is a very simple function
}
\examples{
qDiff<-c(1,2,3,4,5)
qAns<-c(1,1,0,1,0)
testRasch<-new("Rasch", testTakerName="Jim", a=qDiff, y=qAns)
print(testRasch)
}
\seealso{
\code{\link{eap}}
}
\author{
Gangyi Sun
}
