% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRasch.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\alias{EAP}
\title{Prints the name of the test taker and the EAP results}
\usage{
\S4method{print}{Rasch}(x)
}
\arguments{
\item{x}{An object of class Rasch}
}
\value{
A character object
  \item{name}{The subject's name}

A numeric
  \item{ability_estimate}{An estimate of the subject's ability}
}
\description{
Prints the name of the test taker and the EAP results
}
\note{
Should return both the name and the estimated ability level (theta)
}
\examples{

}
\seealso{
\code{\link{EAP}}
}
\author{
Kalen Davison: \email{davison@wustl.edu}
}
