% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRasch.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\title{Printing an EAP and Student Name}
\usage{
\S4method{print}{Rasch}(x)
}
\value{
A list containing the output of the EAP calculation and the student's name
 \item{list(x@name, "EAP" = EAP)}{A list containing the output of the EAP calculation and the student's name}
}
\description{
Print method for Rasch objects which calculates the EAP and prints the result as well as the test taker's name
}
\section{Slots}{

\describe{
\item{\code{x}}{An object of class Rasch}
}}

\examples{

myRasch <- newRasch("Emily", c(3, 4, 12), c(1, 1, 0))
printRasch(myRasch)
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
