% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prints.R
\name{prints}
\alias{prints}
\title{Print}
\usage{
prints(raschObj)
}
\arguments{
\item{name}{The name of the person being analyzed}

\item{EAP}{Expected a Priori value}
}
\value{
A single numeric value that is is the negative LL
}
\description{
Returns the height of the LL for a  given value of lambda
}
\note{
This is a help session file
}
\author{
Dominique Lockett
}
