% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\alias{Prior,ANY-method}
\title{Calculate the prior}
\usage{
Prior(theta)
}
\arguments{
\item{theta}{A proposed value of theta.}
}
\value{
The output containing
 \item{h}{The height of the specified normal curve evaluated at theta}
}
\description{
Calculate the height of the normal curve, given our believe that the prior is 
always going to be a normal distribution
}
\note{
This helps to work in a Bayesian framework
}
\examples{

theta <- 1
Prior(theta)
}
\author{
Luwei Ying
}

