% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorProbFunction.R
\name{priorProbFunction}
\alias{priorProbFunction}
\title{Calculates the prior probablitiy value}
\usage{
priorProbFunction(theta = "numeric")
}
\arguments{
\item{theta}{a numeric value representing a propsed theta value}
}
\value{
The hieght of the distribution at the specified value
}
\description{
Calcualtes the height of a zero mean, standard deviation = 3 normal distribution at specified point
}
\note{
uses dnorm to calculate height
}
\examples{

theta = 3.5
priorProbFunction(theta)
}
\author{
Jonathan Gross <\email{jonathan.gross@wustl.edu}>
}
