% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorRasch.R
\name{priorRasch}
\alias{ANY-method}
\alias{priorRasch}
\alias{priorRasch,}
\title{Prior probability of theta}
\usage{
priorRasch(theta, ...)
}
\arguments{
\item{theta}{A proposed value of theta.}
}
\value{
The prior probability of a given theta.
}
\description{
Calculates the prior probability of theta integrating the probability density function of theta.
}
\examples{
priorRasch(theta = 1)
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\seealso{
\code{\link{Rasch}}
}

