% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\alias{probability,ANY-method}
\title{Finding probability}
\usage{
probability(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An objet of class Rasch}

\item{theta}{A numeric value, the proposed value of ThetaJ}
}
\value{
A list containing:
 \item{Pij}{a vector containing values of Pij for each question}
 \item{PQ}{a vector containing Pij if person j answer question i correctly, and Qij if she answered the question wrongly}
}
\description{
Finds the probability of getting a question right
}
\note{
This is a very simple function
}
\examples{
qDiff<-c(1,2,3,4,5)
qAns<-c(1,1,0,1,0)
testRasch<-new("Rasch", testTakerName="Jim", a=qDiff, y=qAns)
probability(testRasch,3)
}
\seealso{
\code{\link{print}}
}
\author{
Gangyi Sun
}
