% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probFunction.R
\name{probFunction}
\alias{probFunction}
\title{Calculates Probablity of correct response for Rasch Model}
\usage{
probFunction(raschObj = "Rasch", theta = "numeric")
}
\arguments{
\item{raschObj}{an object containing a students name, ability and whether he got a question correct}

\item{theta}{a numeric value representing the ability of a student}
}
\value{
A data frame containing probability of a correct answer, and the probability of the students actual outcome
\item \code{P.vector} a vector containing the probabilities the student gets a each correct
\item \code{PQ} a vector containing the probabilities the student got the output he did
}
\description{
Calcualtes probablitiy of a correct response given theta (ability) and a(difficulty parameter)
}
\note{
Probablity calculated as per equation 1 in documentation
}
\examples{

testRaschObj = new("Rasch", name="Jon",a =c(2,4,3), y =c(1,1,0))
theta = 3.5
probFunction(test1,theta)
}
\author{
Jonathan Gross <\email{jonathan.gross@wustl.edu}>
}
