% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{ProbRasch}
\alias{ProbRasch}
\alias{ProbRasch,ANY-method}
\title{Getting the probabilities}
\usage{
ProbRasch(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of theta.}
}
\value{
The output containing
 \item{P}{A vetor representing the probability of answering the question correctly} 
 \item{PQ}{A vetor representing P if the answer is correct and Q if wrong}
}
\description{
Finds the the probabilities that a test taker get each question right
}
\note{
This function is created for later calculation
}
\examples{

Trump <- new("Rasch", name = "Trump", a = -3:3, y = c(1, 1, 0, 0, 1, 0, 1))
theta <- 1 
ProbRasch(Trump, theta)
}
\author{
Luwei Ying
}
\seealso{
\code{\link{Rasch_class}}
}

