% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch_class.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,Rasch-method}
\title{An object representing a student who has completed an exam}
\description{
Object of class \code{Rasch}
}
\details{
An object of the class ``Rasch" has the following slots:
\itemize{
\item \code{name} The name of the test taker
\item \code{a} A vector of question-item parameters
\item \code{y} A vector of answers 
}
}
\author{
Luwei Ying: \email{luwei.ying@wustl.edu}
}

