% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschClass.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{EAP}
\title{Creates a Rasch Class}
\arguments{
\item{name}{A character object}

\item{difficulty}{A numeric vector}

\item{answer}{A numeric vector}
}
\value{
Creates a Rasch Class
}
\description{
Creates a Rasch Class
}
\note{
Creates Simpson Rasch Class with 3 slots
}
\examples{

}
\seealso{
\code{\link{Prior}}
}
\author{
Kalen Davison: \email{davison@wustl.edu}
}
