% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschEAP.R
\name{raschEAP}
\alias{raschEAP}
\alias{raschEAP,ANY-method}
\title{Expected a posteriori value of theta parameter for the Rasch model.}
\usage{
raschEAP(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{A numeric vector of length one giving the lower limits of
integration; the default value is 6}

\item{upper}{A numeric vector of length one giving the upper limits of
integration; the default value is 6}
}
\value{
A numeric vector of length one giving the expected a posteriori
  value of theta
}
\description{
Determine the expected a posteriori value of theta given the responses we
have observed from a test taker.
}
\details{
The expected a posteriori value (EAP) of theta is calculated by taking the
integral of \eqn{L(\theta_j|y_j)\pi(\theta_j)}, where \eqn{\pi(\theta_j)}
is the prior function for a proposed \eqn{\theta_j}. While the EAP is
defined as the integral from negative to positive infinity, this function
approximates the integral between specified upper and lower limits.
}
\seealso{
\code{\link{Rasch-class}} \code{\link{raschProbability}}
}
\author{
JB Duck-Mayr
}
