% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschInformation.R
\name{raschInformation}
\alias{raschInformation}
\alias{raschInformation,ANY-method}
\title{Fisher Information from questions in a Rasch model for a given test ability.}
\usage{
raschInformation(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{theta}{A numeric vector of length one giving a proposed value for
theta}
}
\value{
A numeric vector of length n (the number of test questions) giving
  Fisher's information for each question for a test taker of ability theta
}
\description{
Determine the Fisher Information for each question in a test for a test
taker with a proposed theta value.
}
\seealso{
\code{\link{Rasch-class}}
}
\author{
JB Duck-Mayr
}
