% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschLikelihood.R
\name{raschLikelihood}
\alias{raschLikelihood}
\alias{raschLikelihood,ANY-method}
\title{Likelihood function for the Rasch model.}
\usage{
raschLikelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A numeric vector of length one giving a proposed value of theta}
}
\value{
A numeric vector of length one giving the likelihood given y & theta
}
\description{
Determine the likelihood of a proposed value of theta given the responses
we have observed from a test taker.
}
\details{
The likelihood of a proposed value of theta given the responses we have
observed from a test taker is given by the product of the probability of
each observed response given the item parameters and the proposed value of
theta. Because it is more numerically stable, this function uses the
exponentiation of the sum of the logged probabilities.
}
\seealso{
\code{\link{Rasch-class}} \code{\link{raschProbability}}
}
\author{
JB Duck-Mayr
}
