% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschModel_EAP.R
\name{raschModel_EAP}
\alias{raschModel_EAP}
\title{Uses the base function integrate() in r to compute the expected a posteriori value for Theta}
\usage{
raschModel_EAP(raschObj, theta, upper = 6, lower = -6)
}
\arguments{
\item{raschObj}{An object containing a students name, question item parameters 'a', and answers 'y'}

\item{theta}{A proposed numeric value for the 'ability' of the respondent}

\item{upper}{The lower limits of integration (default this to 6)}

\item{lower}{The lower limits of integration (default this to -6)}
}
\value{
A estimated value for theta
 \item{object}{An object of class Rasch}
}
\description{
Uses Object from class Rasch
}
\examples{

raschObj <- new(name = "Jason", a = c(1,3,4,3,2), y = c(3,4,3,5,7))
theta <- c(5)
raschModel_EAP(raschObj = raschObj, theta = theta)


}
\seealso{
\code{\link{Rasch}}
}
\author{
Alex Petri
}
