% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschModel_Likelihood.R
\name{raschModel_Likelihood}
\alias{raschModel_Likelihood}
\title{Computes e joint probability of all of the observed responses (P for correct and Q for incorrect) given the guess of theta}
\usage{
raschModel_Likelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{An object containing a students name, question item parameters 'a', and answers 'y'}

\item{theta}{A proposed numeric value for the 'ability' of the respondent}
}
\value{
A calculated likelihood
 \item{object}{An object of class Rasch}
}
\description{
Uses Object from class Rasch
}
\examples{

raschObj <- new(name = "Jason", a = c(1,3,4,3,2), y = c(3,4,3,5,7))
theta <- c(5)
raschModel_Likelihood(raschObj = raschObj, theta = theta)


}
\seealso{
\code{\link{Rasch}}
}
\author{
Alex Petri
}
