% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschModel_Print.R
\docType{methods}
\name{print,ANY-method}
\alias{print,ANY-method}
\title{Printing the subject and their EAP score}
\usage{
\S4method{print}{ANY}(raschObj, theta, upper = 6, lower = -6)
}
\arguments{
\item{raschObj}{An object containing a students name, question item parameters 'a', and answers 'y'}

\item{theta}{A proposed numeric value for the 'ability' of the respondent}

\item{upper}{The lower limits of integration (default this to 6)}

\item{lower}{The lower limits of integration (default this to -6)}
}
\description{
This function prints out the result of of raschModel_EAP and the name of the subject
}
\examples{

raschObj <- new(name = "Jason", a = c(1,3,4,3,2), y = c(3,4,3,5,7))
theta <- c(5)
result <- raschModel_EAP(raschObj = raschObj, theta = theta, upper = 6, lower = -6)
print(result)


}
\seealso{
\code{\link{Rasch}}
}
\author{
Alex Petri
}
