% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschModel_Prior.R
\name{raschModel_Prior}
\alias{raschModel_Prior}
\title{Computes the height of the specified normal curve evaluated at theta with mu = 0, and theta = 3}
\usage{
raschModel_Prior(theta)
}
\arguments{
\item{theta}{A proposed numeric value for the 'ability' of the respondent}
}
\value{
A calculated likelihood
 \item{object}{An object of class Rasch}
}
\description{
Uses Object from class Rasch
}
\examples{

theta <- c(5)
raschModel_Prior(theta = theta)


}
\seealso{
\code{\link{Rasch}}
}
\author{
Alex Petri
}
