% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschModel_Probability.R
\name{raschModel_Probability}
\alias{raschModel_Probability}
\title{Computes the probability for a one parameter item response model, or Rasch Model}
\usage{
raschModel_Probability(raschObj, theta)
}
\arguments{
\item{raschObj}{An object containing a students name, question item parameters 'a', and answers 'y'}

\item{theta}{A proposed numeric value for the 'ability' of the respondent}
}
\value{
A vector of length n that represents P for each question

. A vector of length n that represents P if they got the question correct and Q if wrong
 \item{object}{An object of class Rasch}
}
\description{
Uses Object from class Rasch
}
\examples{

raschObj <- new(name = "Jason", a = c(1,3,4,3,2), y = c(3,4,3,5,7))
theta <- c(5)
raschModel_Probability(raschObj = raschObj, theta = theta)


}
\seealso{
\code{\link{Rasch}}
}
\author{
Alex Petri
}
