% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPrint.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\alias{raschPrint,ANY-method}
\title{EAP print}
\usage{
\S4method{print}{Rasch}(x)
}
\arguments{
\item{raschObj}{An object of class "Rasch." Includes a vector of question difficulty and a corresponding vector indicating whether the student referenced in the object got the question correct (1) or incorrect (0)}
}
\value{
The expected a posteriori value for the student and their name  
 \item{output} The expected a posterior value and the student's name
}
\description{
Prints the EAP value for a certain student
}
\note{
This is a very complicated function function
}
\examples{

a <- c(1,2,3,4,5,6,7,8,9,10)
yJ <- c(1,0,1,1,0,1,1,1,0,1)
Brian <- new("Rasch", name = "Brian", a = a, yJ = yJ)
raschPrint(Brian, lower, upper)
}
\seealso{
\code{\link{raschProbability, raschLikelihood, raschPrior, raschEAP}}
}
\author{
Aaron J. Brezel
}
