% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPrior.R
\name{raschPrior}
\alias{raschPrior}
\alias{raschPrior,ANY-method}
\title{Prior probability of theta parameters in the Rasch model.}
\usage{
raschPrior(theta)
}
\arguments{
\item{theta}{A numeric vector of length one giving a proposed value of theta}
}
\value{
A numeric vector of length one giving the prior probability of the
  proposed value of theta
}
\description{
Determine the prior probability of a proposed value of theta.
}
\seealso{
\code{\link{Rasch-class}}, \code{\link{raschProbability}},
  \code{\link{raschLikelihood}}, \code{\link{raschEAP}}
}
\author{
JB Duck-Mayr
}
