% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschProbability.R
\name{raschProbability}
\alias{raschProbability}
\alias{raschProbability,ANY-method}
\title{Probability function for the Rasch model.}
\usage{
raschProbability(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A numeric vector of length one giving a proposed value of theta}
}
\value{
A list with the following elements:
  \item{questionProbabilities}{A numeric vector of length n that represents
  the probability of a respondent with the proposed theta paramter answering
  correctly for each question}
  \item{answerProbabilities}{A numeric vector of length n that represents
  the probability the respondent \code{rashObj} answered they way they did}
}
\description{
Determine the probabilities a respondent answered a question correctly and
that they gave the observed answer for a proposed theta value.
}
\details{
The probability a respondent answers a question correctly depends both on
the difficulty of the question and the ability of the respondent. For some
proposed value of theta, a parameter for the respondent's ability, this
function returns a list of two numeric vectors, each of length n (the number
of questions); the first vector gives Pij, or the probability the respondent
answered each question j correctly, while the second gives the probability
the respondent answered the way they we observed them responding (i.e. Pij
if we observe a correct answer and Qij = 1 - Pij if we observe an incorrect
answer).
}
\seealso{
\code{\link{Rasch-class}}
}
\author{
JB Duck-Mayr
}
