% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschTestInformationCurve.R
\name{raschTestInformationCurve}
\alias{raschTestInformationCurve}
\alias{raschTestInformationCurve,ANY-method}
\title{Plot a Test Information Curve}
\usage{
raschTestInformationCurve(raschObj, lower = -3, upper = 3)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{lower}{A numeric vector of length one giving the lowest theta value}

\item{upper}{A numeric vector of length one giving the highest theta value}
}
\description{
Plot the sum of Fisher Information for all test questions at a range of
theta values.
}
\seealso{
\code{\link{Rasch-class}}
}
\author{
JB Duck-Mayr
}
