% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIbootCLR}
\alias{RIbootCLR}
\title{Bootstrapped Likelihood Ratio Test}
\usage{
RIbootCLR(dat, iterations = 250, samplesize = 500, cpu = 4)
}
\arguments{
\item{dat}{A dataframe with response data}

\item{iterations}{How many bootstrap samples to run}

\item{samplesize}{How large sample to use in each bootstrap}

\item{cpu}{How many CPU's to use}
}
\description{
Non-parametric bootstrap use of \code{iarm::clr_tests()}.
}
