% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIbootLRT}
\alias{RIbootLRT}
\title{Bootstrapped Likelihood Ratio Test}
\usage{
RIbootLRT(dat, iterations = 1000, samplesize = 500, cpu = 4)
}
\arguments{
\item{dat}{A dataframe with response data}

\item{iterations}{How many bootstrap samples to run}

\item{samplesize}{How large sample to use in each bootstrap}

\item{cpu}{How many CPU's to use}
}
\description{
Non-parametric bootstrap use of \code{iarm::clr_tests()}. Intended for use with
large sample sizes.
}
