% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIbootPCA}
\alias{RIbootPCA}
\title{Parametric bootstrap procedure for PCA of Rasch model residuals}
\usage{
RIbootPCA(
  data,
  iterations = 200,
  cpu = 4,
  rotation = "oblimin",
  maxiter = 5000
)
}
\arguments{
\item{data}{Dataframe with item responses}

\item{iterations}{Number of bootstrap iterations}

\item{cpu}{Number of CPU cores to use}

\item{rotation}{Defaults to "oblimin"}

\item{maxiter}{Maximum number of iterations}
}
\description{
Estimates item and person parameters from data and simulates data fitting the
RM or PCM, then estimates the largest eigenvalue from residuals.
}
\details{
Outputs an object with complete results under \verb{$results} and
percentile values at 95\%, 99\%, 99.5\%, and 99.9\%.
}
