% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIdifFigure}
\alias{RIdifFigure}
\title{Create a DIF line graph, showing groups' PCM item locations}
\usage{
RIdifFigure(dfin, dif.var)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}
}
\description{
Create a DIF line graph, showing groups' PCM item locations
}
