% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIdifTileplot}
\alias{RIdifTileplot}
\title{Check response distribution prior to DIF analysis}
\usage{
RIdifTileplot(data, dif_var)
}
\arguments{
\item{data}{Dataframe with item responses}

\item{dif_var}{DIF variables, ideally a labelled factor}
}
\description{
Outputs a patchwork plot with one \code{RItileplot()} for each level of the DIF
variable. Note that continuous DIF variables, such as age in years, will not
work. A limit has been set at 12 levels of DIF.
}
\details{
Plot labels are automatically set based on the DIF variable being a factor
with labels. You can change the patchwork object title/etc as usual by adding
for instance \code{+ plot_annotation(title = "Whatever you like")}.
}
