% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIestThetas}
\alias{RIestThetas}
\title{Person location estimation}
\usage{
RIestThetas(data, method = "WLE")
}
\arguments{
\item{data}{Dataframe with response data only (no demographics etc), items as columns}

\item{method}{Estimation method (defaults to "WLE")}
}
\description{
Outputs a dataframe of person locations (theta) and measurement error (SEM)
for each person
}
\details{
IMPORTANT: only use with complete response data. If you have missing item responses
\code{RIestThetasCATr()} is recommended instead.

Uses \code{iarm::person_estimates()} to estimate person locations
(thetas) for a dataframe with item data as columns and persons as rows.

Defaults to use WLE estimation (lower bias than MLE, see Warm, 1989) and PCM.

Note: If you want to use a pre-specified set of item parameters, please use
\code{RIestThetasCATr()}.
}
