% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIestThetasCATr}
\alias{RIestThetasCATr}
\title{Person location estimation with parallel processing}
\usage{
RIestThetasCATr(
  data,
  itemParams,
  method = "WL",
  cpu = 4,
  theta_range = c(-10, 10)
)
}
\arguments{
\item{data}{Dataframe with response data only (no demographics etc), items as columns}

\item{itemParams}{Optional item (threshold) location matrix}

\item{method}{Estimation method (defaults to \code{"WL"})}

\item{cpu}{Number of CPUs/cores to utilize (default is 4)}

\item{theta_range}{Range of theta (person location) values}
}
\description{
Yields about 2-3x speed increase when using 4-8 CPU cores.
Requires \code{library(furrr)}
}
\details{
Outputs a vector of person locations, one for each row in the dataframe.

Uses thetaEst function from catR package to estimate person locations
(thetas) for a dataframe with item data as columns and persons as rows.
Defaults to use WL estimation (lower bias than ML, see Warm, 1989).
}
