% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIestThetasOLD}
\alias{RIestThetasOLD}
\title{Person location estimation (\code{catR} version)}
\usage{
RIestThetasOLD(data, itemParams, method = "WL", theta_range = c(-10, 10))
}
\arguments{
\item{data}{Dataframe with response data only (no demographics etc), items as columns}

\item{itemParams}{Optional item (threshold) location matrix}

\item{method}{Estimation method (defaults to "WL")}

\item{theta_range}{Range of theta (person location) values}
}
\description{
Outputs a vector of person locations, one for each row in the dataframe.
}
\details{
Uses \code{thetaEst()} function from \code{catR} package to estimate person locations
(thetas) for a dataframe with item data as columns and persons as rows.
Defaults to use WL estimation (lower bias than ML, see Warm, 1989).

A version for multi-core processing is available as \code{RIestThetasCATr2()}.
}
