% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIgetResidCorG2}
\alias{RIgetResidCorG2}
\title{Get simulation based cutoff values for G^2^ local dependency test}
\usage{
RIgetResidCorG2(data, iterations = 500, cpu = 4)
}
\arguments{
\item{data}{Dataframe with response data}

\item{iterations}{Number of simulation iterations (needed)}

\item{cpu}{Number of CPU cores to use (4 is default)}
}
\description{
Based on Chen & Thissen (2007, DOI: 10.2307/1165285),
uses package \code{mirt} (Chalmers, 2012).
}
\details{
Uses a dataframe with response data to simulate residual correlation values
across n simulations based on estimated item & person locations.

Results include mean, max and difference between the mean and max for each
iteration. Also, 95th, 99th, 99.5th and 99.9th percentile values are
calculated for use with \code{RIresidcorr()} as cutoff value, since the max value
may be spurious and dependent on number of iterations.

Uses multi-core processing. To find how many cores you have on your computer,
use \code{parallel::detectCores()}. Remember to keep 1-2 cores free.
}
