% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIinfitKfold}
\alias{RIinfitKfold}
\title{Cross-validation of conditional item infit}
\usage{
RIinfitKfold(
  data,
  k = 5,
  output = "table",
  sim_iter = 100,
  sim_cpu = 4,
  cutoff = c(0.001, 0.999)
)
}
\arguments{
\item{data}{Dataframe with item responses}

\item{k}{Number of folds to use (default is 5)}

\item{output}{Default \code{table}, options \code{figure}, \code{dataframe}, \code{raw}}

\item{sim_iter}{Number of iterations (depends on sample size)}

\item{sim_cpu}{Number of CPU cores to use}

\item{cutoff}{Truncation at percentile values (see \code{?RIitemfit})}
}
\description{
Creates k random folds using \code{rsample::vfold_cv()};
"V-fold cross-validation (also known as k-fold cross-validation) randomly
splits the data into V groups of roughly equal size (called "folds").
A resample of the analysis data consists of V-1 of the folds",
see \url{https://rsample.tidymodels.org/reference/vfold_cv.html}
}
\details{
Each V-1 dataset is used both for calculating item fit and expected item fit
critical values (using \code{RIgetfit()}). If \code{output = "table"} (default), results
are summarized indicating upper and lower bounds for each item's calculated
infit and simulated expected range. This is based on all V-1 fold combinations.
}
