% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIitemCats}
\alias{RIitemCats}
\title{Individual item category probability plots}
\usage{
RIitemCats(data, items = "all", xlims = c(-6, 6), legend = FALSE)
}
\arguments{
\item{data}{Dataframe with item data only}

\item{items}{Optionally a single item \code{"q4"}, or a vector \code{c("q4","q2")}}

\item{xlims}{Start/end point for x-axis}

\item{legend}{Set to a position such as "left" if desired}
}
\description{
Uses \code{eRm::PCM()} and \code{eRm::plotICC()}.
}
