% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIitemcols}
\alias{RIitemcols}
\title{Item response distribution figure}
\usage{
RIitemcols(
  data,
  ncols = 1,
  labelwrap = 25,
  text_ypos = 6,
  viridis_end = 0.9,
  font = "sans"
)
}
\arguments{
\item{data}{Dataframe/tibble with only item response data coded as integers}

\item{ncols}{How many columns to display, defaults to 1}

\item{labelwrap}{Number of characters to show on each row in item description}

\item{text_ypos}{Position on y axis for the text on each column with n responses}

\item{viridis_end}{If you need to adapt the coloring of text on columns (0.9 is default)}

\item{font}{Choose font family}
}
\description{
A figure showing each item's response distribution and n + \% of responses in each category.
If you use Quarto, you may need to increase the code chunk setting for \code{fig-height} to
something above 5 (default).
}
\details{
Works best with the \code{itemlabels} object set up as described in the package README:
\url{https://pgmj.github.io/easyRasch/#using-the-package}
}
