% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RImissingP}
\alias{RImissingP}
\title{Creates a figure with item missing data descriptives for participants}
\usage{
RImissingP(data, output, N = 10)
}
\arguments{
\item{data}{Dataframe/tibble to create table from}

\item{output}{Optional dataframe with participants with missing data}

\item{N}{Show n participants with most missing data (default = 10)}
}
\description{
The y axis is ordered by most missing responses.
}
\details{
Sample use: \code{RImissingP(data)}
}
