% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIpcmPCA}
\alias{RIpcmPCA}
\title{Fits a Rasch PCM using eRm, and conducts a PCA of residuals to get eigenvalues
using \code{psych::pca()} and reports the top 5 values.}
\usage{
RIpcmPCA(
  dfin,
  output = "table",
  fontsize = 15,
  maxiter = 5000,
  rotation = "oblimin"
)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{output}{Defaults to "table", optional "dataframe"}

\item{fontsize}{Set font size for table}

\item{maxiter}{Maximum number of iterations. Increase if convergence not obtained.}

\item{rotation}{Defaults to "oblimin"}
}
\description{
Proportion of explained variance is calculated using \code{stats::prcomp()}.
}
\details{
Note from \code{?psych::pca}:
The eigenvectors are rescaled by the sqrt of the eigenvalues to produce
the component loadings more typical in factor analysis.

Possible rotations are:
"none", "varimax", "quartimax", "promax", "oblimin", "simplimax", and "cluster".

See Chou & Wang (2010, DOI: 10.1177/0013164410379322) for a simulation study
testing PCA eigenvalues across multiple conditions.
}
