% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIreliability}
\alias{RIreliability}
\title{Reliability metrics}
\usage{
RIreliability(data, conf_int = 0.95, draws = 1000)
}
\arguments{
\item{data}{Dataframe/tibble with only item response data coded as integers}

\item{conf_int}{Desired confidence interval (HDCI)}

\item{draws}{Number of plausible values to generate}
}
\description{
Several metrics are reported, RMU, PSI, and EAP.
}
\details{
RMU, Relative Measurement Uncertainty:
This function uses the \code{TAM} library to estimate the Rasch model using
Marginal Maximum Likelihood and then generates plausible values
(PVs; Mislevy, 1991). The function uses borrowed code, see \code{?RMUreliability}.

The PVs are then used with the RMU method described by Bignardi et al. (2025)
to estimate a mean and confidence interval. The mean is generally similar to
the expected a posteriori (EAP) reliability point estimate (Adams, 2005).
The confidence interval uses the highest continuous density interval (HDCI)
based on the distribution of correlations.

Default setting is to generate 1000 PVs. More are recommended for stable
estimates/CIs. How many more has not been systematically evaluated, but
4000 might be a good starting point. For smaller samples, more PVs is not
very demanding computationally, but be wary of the time it takes to create
thousands of PVs for each respondent in large samples.

PSI, Person Separation Index:
Estimated using functions in the \code{eRm} package, see \code{?eRm::SepRel}.

EAP, Expected a posteriori:
Estimated using functions in the \code{TAM} package, see \code{?TAM::EAPrel}.
}
\examples{
\dontrun{
# comparison of a fully Bayesian Rasch model and PVs
df <- eRm::raschdat1[,1:20] \%>\%
  rownames_to_column("id") \%>\%
  pivot_longer(!id, names_to = "item")

library(brms)
brms_model <- brm(
  value ~ 1 + (1 | item) + (1 | id),
  data    = df,
  chains  = 4,
  cores   = 4,
  family = "bernoulli"
)

posterior_draws <- brms_model \%>\%
  as_draws_df() \%>\%
  dplyr::select(starts_with("r_id")) \%>\%
  t()

RMUreliability(posterior_draws)
RIreliability(eRm::raschdat1[,1:20], draws = 4000)
}

}
\references{
\itemize{
\item Bignardi, G., Kievit, R., & Bürkner, P. C. (2025). A general method for estimating reliability using Bayesian Measurement Uncertainty. PsyArXiv. \href{https://osf.io/preprints/psyarxiv/h54k8}{doi:10.31234/osf.io/h54k8}
\item Mislevy, R. J. (1991). Randomization-Based Inference about Latent Variables from Complex Samples. Psychometrika, 56(2), 177–196. \href{https://doi.org/10.1007/BF02294457}{doi:10.1007/BF02294457}
\item Adams, R. J. (2005). Reliability as a measurement design effect. Studies in Educational Evaluation, 31(2), 162–172. \href{https://doi.org/10.1016/j.stueduc.2005.05.008}{doi:10.1016/j.stueduc.2005.05.008}
}
}
