% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIresidcorrG2}
\alias{RIresidcorrG2}
\title{Correlation matrix of Rasch residuals using G^2^}
\usage{
RIresidcorrG2(data, cutoff, output = "table", ...)
}
\arguments{
\item{data}{Dataframe with item data only}

\item{cutoff}{Relative value above the average of all item residual correlations}

\item{output}{Default HTML table, optional "quarto" or "dataframe"}

\item{...}{Options sent to \code{kbl_rise()} for table output}
}
\description{
Based on Chen & Thissen (2007, DOI: 10.2307/1165285),
uses package \code{mirt} (Chalmers, 2012).
}
\details{
Mandatory option to set relative cutoff-value over the average of all
item residual correlations. It is strongly recommended to use the function
\code{RIgetResidCorG2()} to retrieve an appropriate cutoff value for your data.
}
