% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIrestscore}
\alias{RIrestscore}
\title{Item-restscore}
\usage{
RIrestscore(data, output = "table", sort, p.adj = "BH")
}
\arguments{
\item{data}{A dataframe with response data}

\item{output}{Defaults to a HTML table, optional "quarto" and "dataframe"}

\item{sort}{Optional sorting on absolute difference (descending)}

\item{p.adj}{See \code{?item_restscore()} for options (BH is default)}
}
\description{
A simple wrapper for \code{iarm::item_restscore()}, adding information about
absolute difference in expected and observed values, and item (average)
location.
}
\details{
Please note that item-restscore is likely to produce false positives when sample
size is > 600. It is recommended to use \code{RIbootRestscore()} with large samples.
}
