% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIrestscoreKfold}
\alias{RIrestscoreKfold}
\title{Cross-validation of item-restscore}
\usage{
RIrestscoreKfold(data, k = 5, output = "table")
}
\arguments{
\item{data}{Dataframe with item responses}

\item{k}{Number of folds to use (default is 5)}

\item{output}{Default \code{table}, option \code{dataframe}.}
}
\description{
Creates k random folds using \code{rsample::vfold_cv()};
"V-fold cross-validation (also known as k-fold cross-validation) randomly
splits the data into V groups of roughly equal size (called "folds").
A resample of the analysis data consists of V-1 of the folds",
see \url{https://rsample.tidymodels.org/reference/vfold_cv.html}
}
\details{
Each V-1 dataset is used both for calculating item-restscore and summarises
results based on BH adjusted p-values.
}
