% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIrmPCA}
\alias{RIrmPCA}
\title{Fits the Rasch model for dichotomous data using \code{eRm::RM()}, and
conducts a PCA of residuals to get eigenvalues.}
\usage{
RIrmPCA(
  dfin,
  output = "table",
  fontsize = 15,
  maxiter = 5000,
  rotation = "oblimin"
)
}
\arguments{
\item{output}{Optional "dataframe" or "quarto"}

\item{fontsize}{Set font size}

\item{maxiter}{Maximum number of iterations. Increase if convergence not obtained.}

\item{rotation}{Defaults to "oblimin"}

\item{data}{Dataframe with item data only}
}
\description{
See \code{?RIpcmPCA} for more details.
}
\details{
See Chou & Wang (2010, DOI: 10.1177/0013164410379322) for a simulation study
testing PCA eigenvalues across multiple conditions.
}
