% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIu3poly}
\alias{RIu3poly}
\title{Person fit with U3 for polytomous data}
\usage{
RIu3poly(data)
}
\arguments{
\item{data}{Dataframe with item responses}
}
\description{
This function will estimate person parameters with WLE and item parameters
with CML and use these as input to the \code{PerFit::U3poly()} function. Since
the results from \code{PerFit::U3poly()} are inconsistent, the same analysis is
iterated 100 times, and the median proportion of flagged respondents
is returned.
}
