% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testInf.R
\name{testInf}
\alias{testInf}
\title{Plots the test information for a Rasch model}
\usage{
testInf(raschObj, thetaLim)
}
\arguments{
\item{thetaLim}{A two-element vector specifying the starting and ending values of theta.}

\item{x}{An object of class \code{Rasch}}
}
\value{
A plot of a test information curve
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 exRasch <- new('Rasch', name = 'Me', a = 0, y = 1)
 testInf(exRasch)
 @seealso fisherInf, plotRasch
 @rdname testInf
 @aliases plot,ANY-method
 @export
}
\description{
Plots the test information for a broad range of (user-specified) theta values
from a \code{Rasch} object.
}
