% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eap.R
\name{eap}
\alias{eap}
\alias{eap,}
\alias{ANY-method}
\title{Evaluation of the expected a posteriori}
\usage{
eap(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{an object of class \code{Rasch}}

\item{lower}{a proposed lower bound of integration. The default is -6}

\item{upper}{a proposed upper bound of integration. The default is 6}
}
\value{
the expected a posteriori value
}
\description{
Finds the value of the expected a posteriori given an input object of class \code{Rasch} and a lower and upper bound.
}
\note{
This is a very simple function
}
\examples{

raschobject<-new("Rasch", name="Benjamin", a=c(1,2,3,4,5), y_i=c(0,1,0,1,0))
upper<-5
lower<-0
eap(raschobject, lower, upper)

}
\seealso{
Rasch, likelihood, probability , prio
}
\author{
Benjamin Schneider
}
