% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\title{EAP}
\usage{
EAP(raschObj, lower, upper)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limits of integration (-6)}

\item{upper}{The upper limits of integration (6)}
}
\value{
A value of theta hat specified in the Prior function
}
\description{
Returns the value of theta hat calculated in the Prior function
}
\note{
This is a help session file
}
\author{
Dominique Lockett
}
