% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\title{Calculating Expected a Posteriori Value for Theta}
\usage{
EAP(raschObj, lower = -6, upper = 6)
}
\value{
A numeric representing the likelihood theta is correct
 \item{Top/Bottom}{expected a posteriori value for theta}
}
\description{
Calculating the Likelihood That a Proposed Value of Theta is Correct
}
\section{Slots}{

\describe{
\item{\code{raschObj}}{An object of class Rasch}

\item{\code{lower}}{A numerical input the lower bound defaulted to -6}

\item{\code{upper}}{A numerical input the upper bound defaulted to 6}
}}

\examples{

myRasch <- newRasch("Emily", c(3, 4, 12), c(1, 1, 0))
EAP(myRasch, lower = -6, upper = 6)
}
\seealso{
newRasch
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
