% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\title{Calculate the expected a posteriori value for theta}
\usage{
EAP(raschObj = new("Rasch", name = c("Lim"), a = c(sample(1:10, 100, replace =
  T)), y = c(sample(0:1, 100, replace = T))), lower = -6, upper = 6,
  theta = 1)
}
\arguments{
\item{lower}{the lower limit of integration}

\item{upper}{the upper limit of integration}

\item{theta}{}
}
\value{
A single numeric value that is the negative LL
}
\description{
return the esimtated value of theta
}
\author{
Lim
}
