% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\title{Runs EAP function}
\usage{
EAP(raschObj = "Rasch", lower = "numeric", upper = "numeric")
}
\arguments{
\item{raschObj}{a object of class Rasch}

\item{lower}{lower bound of integral. should be -6}

\item{upper}{upper bound of integral. should be 6}
}
\value{
A numeric
}
\description{
Runs EAP function
}
\note{
I am not exactly sure what math I am suppose to do here. This is my best intepretation
}
\examples{

}
\author{
Jack Ploshnick
}
