% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eap.R
\name{eap}
\alias{eap}
\alias{eap,ANY-method}
\title{Expected a posteriori value}
\usage{
eap(RaschObj, lower = -6, upper = 6)
}
\arguments{
\item{RaschObj}{An object of "Rasch" class}

\item{lower}{Lower limit of integration (default=-6)}

\item{upper}{Upper limit of integration (default=6)}
}
\value{
Expected a posteriori value
}
\description{
The function calculating EAP
}
\note{
This function is just for the exam.
}
\examples{
atesttaker<-new("Rasch", a=c(1:10), yj=rep(c(1,0),5))
eap(atesttaker,-Inf,Inf)
}
\author{
Jaerin Kim
}
