% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\alias{Print}
\title{Rasch Model Ability Estimation}
\usage{
EAP(raschObj = "Rasch", lower = "numeric", upper = "numeric")
}
\arguments{
\item{raschObj}{an object of class Rasch}

\item{theta}{A numeric object}
}
\value{
A numeric
 \item{ability_estimate}{An estimate of the subject's ability}
}
\description{
Estimates the ability of the subject
}
\note{
I do not really understand the math here and am likely to have done this wrong.
}
\examples{

}
\seealso{
\code{\link{raschClass}}
}
\author{
Kalen Davison: \email{davison@wustl.edu}
}
