% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\alias{EAP,}
\alias{ANY-method}
\title{EAP function}
\usage{
EAP(raschObj = "Rasch", lower = "numeric", upper = "numeric")
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{Lower limit of integration}

\item{upper}{Upper limit of integration}
}
\value{
The following value
 \item{theta_hat_j}{The calculated expected a posteriori from Equation 3 on assignment sheet}
}
\description{
EAP function
}
\note{
EAP function
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
