% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eap.R
\name{eap}
\alias{eap}
\title{Calculating the expected a posteriori value for thetaj (our estimate of thetaj)}
\usage{
eap(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limits of integration}

\item{upper}{The upper limits of integration}
}
\value{
A single numeric value that is the value of thetajhat as specified in Equation (3)
}
\description{
Returns the value of thetajhat as specified in Equation (3)
}
\author{
Rohan Gupta<\email{rohangupta@wustl.edu}>
}
