% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eap.R
\name{eap}
\alias{eap}
\title{Calculate the EAP for a Rasch model}
\usage{
eap(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{lower}{The lower limit of integration (default to -6)}

\item{upper}{The upper limit of integration (default to 6)}
}
\value{
A list with one element
 \item{theta}{The value of theta}
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 exRasch <- new('Rasch', name = 'Me', a = 0, y = 1)
 eap(exRasch)
 @seealso likelihood, prior
 @rdname eap
 @aliases eap,ANY-method
 @export
}
\description{
Calculates the expected a posteriori value of an ability parameter (theta) by
integrating over a user-specified domain of values along the posterior function 
generated by a \code{Rasch} object.
}
