% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\docType{methods}
\name{initialize,Rasch-method}
\alias{initialize,Rasch-method}
\title{Expected a posteriori value of theta_j}
\usage{
\S4method{initialize}{Rasch}(.Object, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limits of integration; default is -6}

\item{upper}{The upper limits of integration; default is 6}
}
\value{
The value of theta_j as specified by the scary formula
}
\description{
Estimates the a posteriori value of theta_j using a scary formula
}
\examples{
raschObj<-new(Class = "Rasch", name="Zoe", a=rep(1,3), yj=c(0,1,0))
theta<-1
EAP(raschObj, lower=1, upper=2)

}
\author{
Zoe <\email{zoeang@wustl.edu}>
}
