% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapRasch.R
\name{eapRasch}
\alias{ANY-method}
\alias{eapRasch}
\alias{eapRasch,}
\title{EAP - Expected a posteriori value of theta}
\usage{
eapRasch(raschObj, lower = -6, upper = 6, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limit of the integration}

\item{upper}{The upper limit of the integration}
}
\value{
The estimated a posteriori value of theta
}
\description{
Calculate the expected a posteriori value of theta
}
\examples{
set.seed(1235)
Maggie <- new("Rasch", name = "Maggie", 
a = sample(-3:3, size = 7, replace = TRUE),
y = sample(c(0,1), size = 7, replace = TRUE))
eapRasch(raschObj = Maggie, lower = -10, upper = 10)
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\seealso{
\code{\link{Rasch}}
}

