% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\alias{likelihood,}
\alias{ANY-method}
\title{Likelihood of Rasch}
\usage{
likelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{theta}{a proposed value of the theta paramter of the probability equation}
}
\value{
The output of the likelihood function
}
\description{
Finds the likelihood function for a proposed value of theta and a given object of class \code{Rasch}.
}
\note{
This function produces values for use in the \code{eap} function
}
\examples{

raschobject<-new("Rasch", name="Benjamin", a=c(1,2,3,4,5), y_i=c(0,1,0,1,0))
likelihood(raschobject, 2)

}
\seealso{
\code{\link{Rasch}}, \code{\link{probability}}, \code{\link{prior}}, \code{\link{eap}}, \code{\link{print}}
}
\author{
Benjamin Schneider
}
