% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\alias{likelihood,}
\alias{ANY-method}
\title{Likelihood of Rasch}
\usage{
likelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{A n object of class \code{Rasch}}

\item{theta}{a proposed value of the theta paramter of the probability equation}
}
\value{
The output of the likelihood function
}
\description{
Finds the likelihood function of a proposed value of theta
}
\note{
This is a very simple function
}
\examples{

raschobject<-new("Rasch", name="Benjamin", a=c(1,2,3,4,5), y_i=c(0,1,0,1,0))
likelihood(raschobject, 2)

}
\seealso{
Rasch,
}
\author{
Benjamin Schneider
}
