% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\alias{likelihood,}
\alias{ANY-method}
\title{likelihood function}
\usage{
likelihood(raschObj = "Rasch", theta = "numeric")
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of theta_j}
}
\value{
The following vectors
 \item{lik}{The calculated likelhood from Equation 2 on assignment sheet}
}
\description{
likelihood function
}
\note{
Likelihood function
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
