% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Calculate likelihood for Rasch model}
\usage{
likelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{theta}{A proposed value of theta}
}
\value{
A list with one element
 \item{L}{The calculated likelihood}
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 exRasch <- new('Rasch', name = 'Me', a = 0, y = 1)
 likelihood(exRasch, 0)
 @seealso probability
 @rdname likelihood
 @aliases likelihood,ANY-method
 @export
}
\description{
Calculates the likelihood of a proposed ability parameter (theta) given observed
responses (y) from a \code{Rasch} object.
}
