% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\docType{methods}
\name{initialize,Rasch-method}
\alias{initialize,Rasch-method}
\title{Likelihood of a proposed value of theta_j}
\usage{
\S4method{initialize}{Rasch}(.Object, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed valie of theta_j}
}
\value{
The calculated likelihood of theta_j conditioned on yj
}
\description{
Calculates the likelihood od a proposed value of theta_j givn the responses of the test taker
}
\examples{
raschObj<-new(Class = "Rasch", name="Zoe", a=rep(1,3), yj=c(0,1,0))
theta<-1
likelihood(raschObj, theta)
}
\seealso{
\code{\link{probability}}
}
\author{
Zoe <\email{zoeang@wustl.edu}>
}
