% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRasch.R
\docType{methods}
\name{plot,Rasch-method}
\alias{plot,Rasch-method}
\title{Plots ICCs for Rasch model}
\usage{
\S4method{plot}{Rasch}(x, showTheta = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{Rasch}}

\item{showTheta}{A logical indicating whether the plot should show the theta 
estimate for the user.}
}
\value{
A plot of item characteristic curves
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 exRasch <- new('Rasch', name = 'Me', a = 0, y = 1)
 plot(exRasch)
 @seealso probability
 @rdname plotRasch
 @aliases plot,ANY-method
 @export
}
\description{
Plots the item characteristic curves for a user whose information is contained 
within a \code{Rasch} object.
}
