% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRasch.R
\docType{methods}
\name{plot,Rasch-method}
\alias{ANY-method}
\alias{plot,Rasch-method}
\alias{plotRasch,}
\title{Plot the probability of getting a question correct on possible values of theta.}
\usage{
\S4method{plot}{Rasch}(x, eatTheta = FALSE, minTheta = -5, maxTheta = 5,
  title = NULL)
}
\arguments{
\item{eatTheta}{A logical vector. If TRUE, estimate values for theta are added to the plot.}

\item{minTheta}{A minimum value for theta.}

\item{maxTheta}{A maximum value for theta.}

\item{title}{A character vector with the title defined by the user. If not specified,
the title will be "Item Characteristic Curve for 'name'".}

\item{raschObj}{An object of class Rasch}
}
\value{
A plot with possible values of theta in the x-axis and 
the probability of answer a question correct given the theta in the y-axis.
}
\description{
Plot the probability of getting a question correct on possible values of theta.
}
\examples{
set.seed(1235)
Maggie <- new("Rasch", name = "Maggie", 
a = sample(-3:3, size = 7, replace = TRUE),
y = sample(c(0,1), size = 7, replace = TRUE))
plot(Maggie, eatTheta = TRUE)
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\seealso{
\code{\link{Rasch}}
}

