% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print,}
\alias{ANY-method}
\title{Print function for student name and EAP}
\usage{
print(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{raschObj}{an object of class \code{Rasch}}

\item{lower}{a proposed lower bound of integration. The default is -6}

\item{upper}{a proposed upper bound of integration. The default is 6}
}
\value{
the name of the student and the expected a posteriori value
}
\description{
Lists the name of the student taking the test and finds the value of the expected a posteriori given an input object of class \code{Rasch} and a lower and upper bound.
}
\note{
This function is a slight upgrade of the \code{eap} function in that it also provides the name of the student.
}
\examples{

raschobject<-new("Rasch", name="Benjamin", a=c(1,2,3,4,5), y_i=c(0,1,0,1,0))
upper<-5
lower<-0
print(raschobject, lower, upper)

}
\seealso{
\code{\link{Rasch}}, \code{\link{likelihood}}, \code{\link{prior}}, \code{\link{eap}}, \code{\link{probability}}
}
\author{
Benjamin Schneider
}
