% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\title{Print}
\usage{
\S4method{print}{Rasch}(x, y, startandend, Rule, ...)
}
\arguments{
\item{x}{}
}
\value{
print the name of test taker and EAP value of theta
}
\description{
print the name of test taker and EAP value of theta
}
\author{
Lim
}
