% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\title{Runs Print function}
\usage{
\S4method{print}{Rasch}(x)
}
\arguments{
\item{x}{object of class Rasch}
}
\value{
Print of Name, integral value, integral error
}
\description{
Runs Print function
}
\note{
Very simple print function
}
\examples{

}
\author{
Jack Ploshnick
}
