% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,ANY-method}
\alias{print,ANY-method}
\alias{print}
\title{Print Method for Rasch Object}
\usage{
\S4method{print}{ANY}(raschObj = "Rasch")
}
\arguments{
\item{raschObj}{an object containing a students name, ability and whether he got a question correct}
}
\description{
Prints name and A posteriori estimate for theta
}
\note{
Calls EAPFunction
}
\examples{

testRaschObj = new("Rasch", name="Jon",a =c(2,4,3), y =c(1,1,0))
print(testRaschObj)
}
\author{
Jonathan Gross <\email{jonathan.gross@wustl.edu}>
}
