% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\alias{print}
\title{print function}
\usage{
\S4method{print}{Rasch}(x)
}
\arguments{
\item{x}{Object of class Rasch}
}
\value{
Test-taker student's name and the EAP results
}
\description{
print function
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
