% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior,}
\alias{ANY-method}
\title{Prior of a given theta}
\usage{
prior(theta)
}
\arguments{
\item{theta}{a proposed value of the theta paramter of the probability equation}
}
\value{
The height of a specified normal curve evaluated at theta when mean=0 and standard deviation=3
}
\description{
Finds the height of a specified normal curve evaluated at a given theta.
}
\note{
This output is used in the \code{eap} function
}
\examples{

theta<-2
prior(theta)

}
\seealso{
\code{\link{Rasch}}, \code{\link{likelihood}}, \code{\link{probability}}, \code{\link{eap}}, \code{\link{print}}
}
\author{
Benjamin Schneider
}
