% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior,}
\alias{ANY-method}
\title{Prior of a given theta}
\usage{
prior(theta)
}
\arguments{
\item{theta}{a proposed value of the theta paramter of the probability equation}
}
\value{
The height of a specified normal curve evaluated at theta when mean=0 and standard deviation=3
}
\description{
Finds the height of a specified normal curve evaluated at theta
}
\note{
This is a very simple function
}
\examples{

theta<-2
prior(theta)

}
\seealso{
Rasch, likelihood, probability
}
\author{
Benjamin Schneider
}
