% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\title{Calculating Prior}
\usage{
Prior(theta)
}
\value{
A numeric representing the height of the normal curve
 \item{dnorm(theta, mean = 0, sd = 3)}{The normal distribution of theta with mean of 0 and standard deviation of 3}
}
\description{
Calculating Normal Distribution of Theta
}
\section{Slots}{

\describe{
\item{\code{theta}}{A numerical input for theta}
}}

\examples{

Prior(4)
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
