% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior,ANY-method}
\title{Finding prior}
\usage{
prior(theta, ...)
}
\arguments{
\item{theta}{A numeric value, the proposed value of ThetaJ}
}
\value{
\item{output}{the height of the specified normal curve evaluated at thetaJ}
}
\description{
Finds the height of the specified normal curve evaluated at thetaJ
}
\note{
This is a very simple function
}
\examples{
prior(3)
}
\seealso{
\code{\link{print}}
}
\author{
Gangyi Sun
}
