% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\title{calculating prior value for Bayesian framework}
\usage{
Prior(theta = 1)
}
\arguments{
\item{theta}{A proposed value of ??_j}
}
\value{
A single numeric value that is return the height of the specified normal curve evaluated at ??_j
}
\description{
return the height of the specified normal curve evaluated at ??_j
}
\author{
Lim
}
