% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\alias{Prior,ANY-method}
\title{A function to create a prior}
\usage{
Prior(theta, ...)
}
\arguments{
\item{theta}{A proposed value of beta}
}
\description{
Calculate the height of the normal curve for a proposed value of theta
}
\examples{
Prior(theta = 1)
Prior(theta = 10)
Prior(theta = 5)

}
\author{
Hyunjoo Oh: \email{hyunjoo.oh@wustl.edu}
}
\seealso{
\code{\link{Probability}}, \code{\link{Likelihood}}
}

