% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\title{Runs Prior function}
\usage{
Prior(theta = "numeric")
}
\arguments{
\item{theta}{numeric designating difficulty}
}
\value{
dnorm mean=0, sd=3 of theta
}
\description{
Runs Prior function
}
\note{
simply calculates dnorm
}
\examples{

}
\author{
Jack Ploshnick
}
