% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior,ANY-method}
\title{Prior}
\usage{
prior(theta)
}
\arguments{
\item{theta}{A proposed value of theta}
}
\value{
Prior value given the theta
}
\description{
The function calculating prior per given theta
}
\note{
This function is just for the exam.
}
\examples{
prior(2)
}
\author{
Jaerin Kim
}
