% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\alias{Print}
\title{Rasch Model Prior}
\usage{
Prior(theta = "numeric")
}
\arguments{
\item{theta}{A numeric object}
}
\value{
A numeric
 \item{height}{The height of the normal curve at a proposed theta}
}
\description{
Calculates the height of the normal curve given a theta value
}
\note{
This is a very simple function
}
\examples{

theta = 1.2
Prior(theta)
}
\seealso{
\code{\link{Likelihood}}
}
\author{
Kalen Davison: \email{davison@wustl.edu}
}
