% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior,}
\alias{ANY-method}
\title{prior function}
\usage{
prior(theta = "numeric")
}
\arguments{
\item{theta}{A proposed value of theta_j}
}
\value{
The following value
 \item{height}{The height of the specified normal curve - mean = 0, standard dev = 3, evaluated at theta_j}
}
\description{
prior function
}
\note{
Prior function
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
