% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Calculating the height of the normal curve for a proposed value of theta}
\usage{
prior(theta)
}
\arguments{
\item{theta}{A proposed value of thetaj}
}
\value{
A single numeric value that is the height of the specified normal curve evaluated at thetaj
}
\description{
Returns the height of the specified normal curve evaluated at thetaj
}
\author{
Rohan Gupta<\email{rohangupta@wustl.edu}>
}
