% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Calculate prior for Rasch model}
\usage{
prior(theta, ...)
}
\arguments{
\item{theta}{A proposed value of theta}
}
\value{
A list with one element
 \item{pi}{The height of the specified normal curve evaluated at theta}
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 prior(0)
 @seealso EAP
 @rdname prior
 @aliases prior,ANY-method
 @export
}
\description{
Calculates the prior probability of a proposed ability parameter (theta) for 
a \code{Rasch} object. The prior is always normally distributed with mean 0 and
standard deviation 3.
}
