% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{methods}
\name{initialize,numeric-method}
\alias{initialize,numeric-method}
\title{Estimate theta_j using Bayesian prior}
\usage{
\S4method{initialize}{numeric}(.Object, ...)
}
\arguments{
\item{theta}{A proposed valie of theta_j}
}
\value{
The height of the specified nromal curve evaluated at theta_j
}
\description{
Estimates theta_j using Bayesian prior to get a more reasonable estimate of the parameter; this wull use the normal curve with mean=0 and sd=3
}
\author{
Zoe <\email{zoeang@wustl.edu}>
}
