% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorRasch.R
\name{priorRasch}
\alias{ANY-method}
\alias{priorRasch}
\alias{priorRasch,}
\title{Calculates the prior probability of theta}
\usage{
priorRasch(raschObj, theta, ...)
}
\arguments{
\item{theta}{A proposed value of \theta[j]}
}
\value{
The height of the specified normal curve evaluated at the theta
}
\description{
Calculates the prior probability of theta
}
\examples{
priorRasch(theta = 1)
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\seealso{
\code{\link{Rasch}}
}

