% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\alias{probability,}
\alias{ANY-method}
\title{Probability of Rasch}
\usage{
probability(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{theta}{a proposed value of the theta paramter of the probability equation}
}
\value{
A vector of length n for \code{P_ij} for each equation and a specialized vector \code{PQ_ij} containing P_ij and Q_ij values depending on correct or incorrect answers.
}
\description{
Finds the probability of a student getting certain exam questions correctly given a parameter theta.
}
\note{
This function produces vectors for use in finding the likelihood function.
}
\examples{

raschobject<-new("Rasch", name="Benjamin", a=c(1,2,3,4,5), y_i=c(0,1,0,1,0))
probability(raschobject, 2)

}
\seealso{
\code{\link{Rasch}}, \code{\link{likelihood}}, \code{\link{prior}}, \code{\link{eap}}, \code{\link{print}}
}
\author{
Benjamin Schneider
}
