% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\alias{probability,}
\alias{ANY-method}
\title{Probability of Rasch}
\usage{
probability(raschObj, theta)
}
\arguments{
\item{raschObj}{A n object of class \code{Rasch}}

\item{theta}{a proposed value of the theta paramter of the probability equation}
}
\value{
A vector of length n for P_ij for each equation and a specialized vector containing P_ij and Q_ij values depending on correct or incorrect answers
}
\description{
Finds the probability of a student getting certain exam questions correctly given a parameter
}
\note{
This is a very simple function
}
\examples{

raschobject<-new("Rasch", name="Benjamin", a=c(1,2,3,4,5), y_i=c(0,1,0,1,0))
probability(raschobject, 2)

}
\seealso{
Rasch, likelihood, prior
}
\author{
Benjamin Schneider
}
