% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{probability}
\alias{probability}
\title{Probability}
\usage{
probability(raschObj, theta)
}
\arguments{
\item{raschObj}{An objecto of class Rasch}

\item{theta}{A proposed value representing a person's ability}
}
\value{
A vector of P and a vector of PQ
}
\description{
Returns probability that a student gets a question right given their ability, and the difficulty of the question
}
\note{
This is a help session file
}
\examples{
dom<-new("Rasch",name="dom", a=c(4,2,3,4,5,6,7,8,9), y=c(1,1,1,1,1,0,1,1,1))
probability(dom,.2)

}
\author{
Dominique Lockett
}
