% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{Probability}
\alias{Probability}
\title{calculating Probability of getting right and wrong by Rasch model}
\usage{
Probability(raschObj = new("Rasch", name = c("Lim"), a = c(sample(1:10, 100,
  replace = T)), y = c(sample(0:1, 100, replace = T))), theta = 1)
}
\arguments{
\item{raschObj}{A numeric value of Rasch values including a and y.}

\item{theta}{A numeric vector containing my poisson data}
}
\value{
A vector of length n that represents Pij for each question and
         A vector of length n that represents Pij
}
\description{
return A vector of length n that represents Pij for each question and
A vector of length n that represents Pij.
}
\author{
Lim
}
