% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{Probability}
\alias{Probability}
\alias{Print}
\title{Rasch Model Probability}
\usage{
Probability(raschObj = "Rasch", theta = "numeric")
}
\arguments{
\item{raschObj}{an object of class Rasch}

\item{theta}{A numeric object}
}
\value{
A list of two vectors
 \item{P}{The probability of each person answering the question correctly}
 \item{PQ}{The probability of each person answering correctly if they got it right, or the probability of each person answering incorrectly if they got it wrong}
}
\description{
Calculates probability using the Rasch formula
}
\note{
Theta value can be manually chosen.
}
\examples{

testRasch = new("Rasch", name = "Benny", difficulty = 1:10, answers = sample(0:1, 10, replace=TRUE))
Probability(raschObj = testRasch, theta = .5)
}
\seealso{
\code{\link{raschClass}}
}
\author{
Kalen Davison: \email{davison@wustl.edu}
}
