% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\alias{probability,}
\alias{ANY-method}
\title{probability function}
\usage{
probability(raschObj = "Rasch", theta = "numeric")
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of theta_j}
}
\value{
The following vectors
 \item{P_ij}{a vector of length n representing P_ij for each question}
 \item{PQ}{a vector of length n representing P_ij if question was answered correctly, otherwise Q_ij}
}
\description{
probability function
}
\note{
Probability function
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
