% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Calculating the probability of answering the question correctly}
\usage{
probability(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of thetaj}
}
\value{
A list with the elements
 \item{P}{A vector of length n that represents Pij for each question}
 \item{PQ}{A vector of length n that represents Pij if they got the question correct and Qij if they got the question wrong}
}
\description{
Returns two vectors of length n that represent Pij for each question and Pij if they got the question correct and Qij if they got the question wrong
}
\author{
Rohan Gupta<\email{rohangupta@wustl.edu}>
}
