% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Calculate probabilities for Rasch model}
\usage{
probability(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class \code{Rasch}}

\item{theta}{A proposed value of theta}
}
\value{
A list with the elements
 \item{P}{A vector of length n that represents P for each question}
 \item{Mixed}{A vector of length n that represents P if the question is answered
 correctly and Q otherwise}
 
 @author Ryden W. Butler: \email{r.butler@wustl.edu}
 @examples 
 
 exRasch <- new('Rasch', name = 'Me', a = 0, y = 1)
 probability(exRasch, 0)
 @seealso likelihood
 @rdname probability
 @aliases probability,ANY-method
 @export
}
\description{
Calculates the probabilities of correct responses (P) and incorrect responses (Q)
for a \code{Rasch} object.
}
