% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\docType{methods}
\name{initialize,Rasch-method}
\alias{initialize,Rasch-method}
\title{A probability function for Pij}
\usage{
\S4method{initialize}{Rasch}(.Object, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed valie of theta_j}
}
\value{
A vector of length n that represents Pij for each question
}
\description{
Calculates the probability of Pij, the output of a Rasch model
}
\examples{
raschObj<-new(Class = "Rasch", name="Zoe", a=rep(1,3), yj=c(0,1,0))
theta<-1
probability(raschObj, theta)
}
\seealso{
\code{\link{Rasch}} for object class
}
\author{
Zoe <\email{zoeang@wustl.edu}>
}
