% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{A probability function for Pij}
\usage{
probability(object = raschObj, theta)
}
\arguments{
\item{theta}{A proposed valie of theta_j}

\item{raschObj}{An object of class Rasch}
}
\value{
A vector of length n that represents Pij for each question
}
\description{
Calculates the probability of Pij, the output of a Rasch model
}
\author{
Zoe <\email{zoeang@wustl.edu}>
}
