% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probRasch.R
\name{probRasch}
\alias{ANY-method}
\alias{probRasch}
\alias{probRasch,}
\title{Probabiltiy Rasch}
\usage{
probRasch(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed value of theta}
}
\value{
A list contains
\itemize{
 \item A vector of length n that represents P_{ij} for each question
 \item A vector of length n that represents P_{ij} if the question was 
 answered correct and Q_{ij} if the question was answered wrong
 }
}
\description{
Calculate the probabiltiy for a object Rasch
}
\examples{
set.seed(1235)
Maggie <- new("Rasch", name = "Maggie", 
a = sample(-3:3, size = 7, replace = TRUE),
y = sample(c(0,1), size = 7, replace = TRUE))
probRasch(raschObj = Maggie, theta = 2)
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}
\seealso{
\code{\link{Rasch}}
}

