% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{A Rasch Model Object}
\description{
Object of class \code{Rasch} are created for raschModel functions
}
\details{
An object of the class "Rasch" has the following slots:
\itemize{
\item \code{name} A vector of the x values of the input
\item \code{a} A vector of question-item parameters
\item \code{y} A vector of answers for the respondent
}
}
\author{
Alex Petri: \email{alexpetri@wustl.edu}
}
