% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,}
\alias{Rasch-method}
\title{Rasch Object}
\description{
Object of class \code{Rasch} is an input in package \code{easyRasch}
}
\author{
Benjamin R. Schneider: \email{benjamin.schneider@wustl.edu}
}
