% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,}
\alias{Rasch-method}
\title{Rasch Object}
\description{
Object of class \code{Rasch} is an input in multiple functions in the \code{easyRasch} package
}
\details{
An object of the class `Rasch' has the following slots:
\itemize{
\item \code{name} The name of the test taker
\item \code{a}  a vector of question item parameters
\item \code{y_i} a vector of answered for the respondent y_j}
}
\author{
Benjamin R. Schneider: \email{benjamin.schneider@wustl.edu}
}
