% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,Rasch-method}
\title{Rasch}
\description{
Objects of class Rasch come from  several functions
}
\details{
An object of the class `Rasch' has the following slots:
\itemize{
\item \code{name} The name of a test-taker
\item \code{a} A set of question-item parameters
\item \code{y} A vector of answers for the respondent

}
}
\author{
Dominique M. Lockett: \email{dlockett@wustl.edu}
}
