% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch-class.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,Rasch-method}
\title{An object with a test respondent's name and responses, and item parameters.}
\description{
Object of class \code{Rasch}
}
\section{Slots}{

\describe{
\item{\code{name}}{A character vector giving the name of the test taker}

\item{\code{a}}{A numeric vector giving the question-item parameters}

\item{\code{y}}{A numeric vector of answers for the respondent; a value of zero
indicates the respondent did not answer correctly and a value of one
indicates the respondent did answer correctly}
}}

\author{
JB Duck-Mayr
}
