% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,Rasch-method}
\title{A Rasch object}
\description{
Object of class \code{Rasch} is created by the Rasch class' initialization function
}
\details{
An object of the class `Rasch' has the following slots:
\itemize{
\item \code{testTakerName} a character of the test taker's name
\item \code{a} a vector of question-item parameters ('difficulty')
\item \code{y} a vector of answers for the respondent (=1 if answered correctly, =0 if wrong.)
}
}
\author{
Gangyi Sun: \email{gangyi.sun@wustl.edu}
}
