% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,}
\alias{Rasch}
\alias{model}
\title{Rasch model class}
\description{
Object of class Rasch is from the input of the Rasch model
}
\details{
An object of the class `Rasch' has the following slots:
\itemize{
\item \code{name} The name of the test taker
\item \code{a} A vector of question-item parameters. Difficulty parameters
\item \code{y} A vector of answers for the respondent. Observed person's answers 
}
}
\author{
Lim
}
