% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{Rasch}
\alias{Rasch,ANY-method}
\title{Rasch Class}
\arguments{
\item{name}{The test taker's name}

\item{a}{A vector of question-item parameters}

\item{yj}{A vector of answers for the respondent}
}
\description{
Defines class Rasch
}
\note{
This class is just for the exam.
}
\examples{
atesttaker<-new("Rasch", a=c(1:10), yj=rep(c(1,0),5))
atesttaker

}
\author{
Jaerin Kim
}
