% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize}
\title{A Rasch Object}
\description{
Object of class \code{Rasch} is useed by the \code{probFunction}
}
\details{
An object of the class `Rasch' has the following slots:
\itemize{
\item \code{name} The name of the student
\item \code{a} A vector containing difficulty parameters 
\item \code{y} A vector containing responses of student (1 = correct, 0 =wrong)
}
}
\author{
Jonathan Gross: \email{jonathan.gross@wustl.edu}
}
