% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,Rasch-method}
\title{A student who has taken a test}
\description{
An object of the class 'Rasch' has the following slots:
\itemize{
\item \code{name} The name of the test taker
\item \code{a} A vector of question-item difficulty paramaters
\item \code{y} A vector of answers for the respondent
}
}
\author{
Noah Bardash: \email{noah.bardash@wustl.edu}
}
