% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch-Class.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{Squares-Rasch}
\alias{initialize,Rasch-method}
\title{Generate class Rasch}
\description{
An object of the class Rasch must contain:
\itemize{
\item \code{name} The name of the test taker
\item \code{a} A vector of the question-item paramenters a (difficulty of the question) 
\item \code{y} A vector of answers for the respondent \eqn{y_j}
}
}
\author{
Patrick C. Silva: \email{pcunhasilva@wustl.edu}
}

