% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{initialize,Rasch-method}
\title{A student who has completed an exam}
\description{
Object of class Rasch
}
\details{
An object of the class 'Rasch' has the following slots:
\itemize{
\item \code{name} The name of the test taker
\item \code{a} A vector of question-item parameters
\item \code{yj} A vector of answers for the respondent
}
}
\author{
Rohan Gupta: \email{rohangupta@wustl.edu}
}
