% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch-Class.R
\name{examineRasch}
\alias{examineRasch}
\title{An object for calculating simple Rasch models}
\usage{
examineRasch(object)
}
\description{
An object of class \code{Rasch} for use in estimating the ability parameter (theta)
of test takers. For use with \code{makeRasch}.
}
\details{
An object of class \code{Rasch} has the following slots:
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the test taker}

\item{\code{a}}{A vector of question-item parameters}

\item{\code{y}}{A vector of answers for the respondent}
}}

\author{
Ryden W. Butler: \email{r.butler@wustl.edu}
}
