% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{An object of class Rasch}
\description{
Object of class Rasch, which relates to the Rasch model
}
\details{
An object of class Rasch has the following slots:
\itemize{
\item \code{name} The name of the test taker
\item \code{a} A vector of question-item parameters
\item \code{yj} A vector of answers for the respondent
}
}
\author{
Zoe <\email{zoeang@wustl.edu}>
}
