% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschClass.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{A rasch}
\description{
Object of class \code{Rasch} are created by the \code{integrateIt} functions
}
\details{
An object of the class "Trapezoid" or "Simpsons" has the following slots:
\itemize{
\item \code{name} A character string indicating the name of the student
\item \code{a} A vector of difficulty values for each test question
\item \code{yJ} A vector off corresponding binary values indicating a correct (1) or incorrect (0) answer from the student
}
}
\author{
Aaron J. Brezel: \email{aaronbrezel@wustl.edu}
}
