% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschEAP.R
\name{raschEAP}
\alias{raschEAP}
\alias{raschEAP,ANY-method}
\title{Calculate the expected a posteriori value for theta}
\usage{
raschEAP(raschObj, lower, upper)
}
\arguments{
\item{raschObj}{An object of class "Rasch." Includes a vector of question difficulty and a corresponding vector indicating whether the student referenced in the object got the question correct (1) or incorrect (0)}

\item{lower}{A numeric value indicating the lowerbound of the comming integration}

\item{upper}{A numeric value indicating the upperbound of the comming integration}
}
\value{
The expected a posteriori value for theta  
 \item{output}
}
\description{
Takes and object of class "Rasch", and uses it to calculate the expected a posteriori value
}
\note{
This is a very complicated function function
}
\examples{

a <- c(1,2,3,4,5,6,7,8,9,10)
yJ <- c(1,0,1,1,0,1,1,1,0,1)
Brian <- new("Rasch", name = "Brian", a = a, yJ = yJ)
lower <- -6
upper <- 6
raschEAP(Brian, lower, upper)
}
\seealso{
\code{\link{raschProbability, raschLikelihood, raschPrior}}
}
\author{
Aaron J. Brezel
}
