% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschLikelihood.R
\name{raschLikelihood}
\alias{raschLikelihood}
\alias{raschLikelihood,ANY-method}
\title{Calculated likelihood of a proposed value of theta}
\usage{
raschLikelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class "Rasch." Includes a vector of question difficulty and a corresponding vector indicating whether the student referenced in the object got the question correct (1) or incorrect (0)}

\item{theta}{A numeric value indicating the skill of the student taking the test}
}
\value{
The likelihood of a proposed value of theta  
 \item{output}{The product sum calculated by multiplying the results of the second output of raschProbability all together}
}
\description{
Takes an object of class Rasch (which includes the difficulty of each question) and a theta ("skill of the student") value
and uses them to calculate the likelihood of a proposed value of theta
}
\note{
This is a very complicated function function
}
\examples{

a <- c(1,2,3,4,5,6,7,8,9,10)
yJ <- c(1,0,1,1,0,1,1,1,0,1)
Brian <- new("Rasch", name = "Brian", a = a, yJ = yJ)
theta = 5
raschLikelihood(Brian, theta)
}
\seealso{
\code{\link{raschProbability}}
}
\author{
Aaron J. Brezel
}
