% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPrior.R
\name{raschPrior}
\alias{raschPrior}
\alias{raschPrior,ANY-method}
\title{Calculate the height of a normal curve at the specified theta value}
\usage{
raschPrior(theta)
}
\arguments{
\item{theta}{A numeric value indicating the skill of the student taking the test}
}
\value{
\item{output}{The height of a normal curve with a mean of 0 and a standard deviation of 3 at the theta (student skill) value}
}
\description{
Takes a theta ("skill of the student") value and uses to calculate the height of the normal distribution curve at that value
}
\note{
This is a very complicated function function
}
\examples{

theta = 5
raschPrior(theta)
}
\author{
Aaron J. Brezel
}
