% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschProbability.R
\name{raschProbability}
\alias{raschProbability}
\title{Calculated probability of answering the question correctly}
\usage{
raschProbability(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class "Rasch." Includes a vector of question difficulty and a corresponding vector indicating whether the student referenced in the object got the question correct (1) or incorrect (0)}

\item{theta}{A numeric value of corresponding y values with the same dimensionality as \code{x}.}
}
\value{
An output list. The first slot contains the probability of the student answering each question correctly given the question's difficulty and their skill level (theta). The second slot contains a similar list. Instead of only being the probability of getting a question right, in the slots where the student got the question wrong, the probability of correct value is replaced with the probability of getting the question wrong.  
 \item{output}{The output list}
}
\description{
Takes an object of class Rasch (which includes the difficulty of each question) and a theta ("skill of the student") value
and uses them to calculate the probability of answering a question correctly.
}
\note{
This is a very complicated function function
}
\examples{

a <- c(1,2,3,4,5,6,7,8,9,10)
yJ <- c(1,0,1,1,0,1,1,1,0,1)
Brian <- new("Rasch", name = "Brian", a = a, yJ = yJ)
theta = 5
raschProbability(Brian, 5)
}
\author{
Aaron J. Brezel
}
