% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIbootRestscore}
\alias{RIbootRestscore}
\title{Item-restscore bootstrapped}
\usage{
RIbootRestscore(
  dat,
  iterations = 200,
  samplesize = 600,
  cpu = 4,
  output = "table",
  model
)
}
\arguments{
\item{iterations}{How many bootstrap samples to run}

\item{samplesize}{How large sample to use in each bootstrap}

\item{cpu}{How many CPU's to use}

\item{output}{Optional "dataframe", or "quarto" for \code{knitr::kable()} output}

\item{data}{Dataframe with only response data, with 0 as lowest response}
}
\description{
The text below is based on an ongoing simulation study, as of 2024-10-22.
}
\details{
Item-restscore will often indicate false positives (item misfit when it is not
misfitting) if the sample size is above 400 and there is one truly misfitting
item in the data. If there is more than one misfitting item, false positives
can occur at sample sizes of n = 150-200.

Conversely, when sample size is below n = 800, the detection rate of truly
misfitting items is below 90\%, particularly if misfitting items have
location > 1.5 logits from the sample mean.

Thus, if one has a large dataset it may be useful to be able to use
non-parametric bootstrapping with replacement to get a more nuanced view of
the probability of items actually being misfit.
}
