% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIinfitKfoldPlot}
\alias{RIinfitKfoldPlot}
\title{Plot k-fold infit based on raw output}
\usage{
RIinfitKfoldPlot(kfold)
}
\arguments{
\item{kfold}{Object with "raw" output from \code{RIinfitKfold()}}
}
\description{
Outputs a figure showing highest and lowest expected values based on all
simulations and cross-validation folds from an object created with
\code{RIinfitKfold(data, output = "raw")}.
}
