% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch.R
\name{RIrmPCA}
\alias{RIrmPCA}
\title{Fits the Rasch model for dichotomous data using \code{eRm::RM()}, and
conducts a PCA of residuals to get eigenvalues.}
\usage{
RIrmPCA(data, output = "table", fontsize = 15)
}
\arguments{
\item{data}{Dataframe with item data only}

\item{output}{Optional "dataframe" or "quarto"}

\item{fontsize}{Set font size}
}
\description{
See \code{?RIpcmPCA} for more details.
}
\details{
See Chou & Wang (2010, DOI: 10.1177/0013164410379322) for a simulation study
testing PCA eigenvalues across multiple conditions.
}
